added selector and move reducer
