+++ frontend/src/pages/Market/components/ProductItem/ProductItem.module.css
@@ -1,18 +1,15 @@
 .card {
-    width: 320px;
-    height: 260px;
     margin: 5px;
 }
 
 .cardWrap {
     display: flex;
     justify-content: space-between;
-    flex-direction: column;
     align-items: center;
     padding-top: 24px;
     padding-bottom: 16px;
     border: 1px solid black;
-    border-radius: 5%;
+    border-radius: 5px;
 }
 
 .cardImage {
