+++ frontend/src/pages/Market/components/ProductList/ProductList.module.css
@@ -2,6 +2,6 @@
     width: 97%;
     margin: 0 auto;
     display: flex;
-    align-items: center;
+    flex-direction: column;
     justify-content: space-evenly;
 }
