+++ frontend/src/pages/Market/index.ts
@@ -1,1 +1,2 @@
 export { Market } from "./Market";
+export { initialState, reducer } from "./Market.reducer";
