--- frontend/src/pages/Market/reducer/index.ts
@@ -1,1 +1,0 @@
-export { initialState, reducer } from "./Market.reducer";
